refactor(scheduler-core): extract logic library to mintplayer-web-components
scheduler-core is a pure logic library (models, services, utils — no
Lit element, no DOM). git mv from libs/mintplayer-ng-bootstrap/
web-components/scheduler-core/ to libs/mintplayer-web-components/
scheduler-core/. Legacy ng-package.js secondary entry removed.

Since there's no element class, no React/Vue wrapper is shipped. The
JS API is consumed directly via @mintplayer/web-components/scheduler-
core (the path-mapped subpath import). Both the upcoming scheduler WC
extraction (commit 4) and any downstream consumers will hit the new
path.

Importer rewires (22 files total):
- The 17 internal files inside the still-on-master scheduler WC — they
  reference scheduler-core types throughout (date utilities, resource
  models, drag state, position service). Updated here so the build
  stays green between scheduler-core's move and scheduler's own move
  in the next commit. Scheduler's own files will be moved in commit 4
  via git-mv — their new-location imports are already on the shared
  path.
- libs/mintplayer-ng-bootstrap/scheduler/src/components/scheduler/scheduler.component.ts (Angular wrapper)
- libs/mintplayer-ng-bootstrap/scheduler/src/index.ts (Angular barrel)
- apps/ng-bootstrap-demo/src/app/pages/enterprise/scheduler/scheduler.component.ts (demo)
