+++ libs/mintplayer-ng-bootstrap/web-components/scheduler/src/drag/drag-preview.ts
@@ -3,7 +3,7 @@ import {
   PreviewEvent,
   SchedulerEvent,
   TimeSlot,
-} from '@mintplayer/ng-bootstrap/web-components/scheduler-core';
+} from '@mintplayer/web-components/scheduler-core';
 import { DragConfig, DEFAULT_DRAG_CONFIG } from './drag-types';
 
 /**
