+++ libs/mintplayer-ng-bootstrap/web-components/scheduler/src/drag/drag-state-machine.spec.ts
@@ -1,6 +1,6 @@
 import { DragStateMachine } from './drag-state-machine';
 import { DragMachineEvent, PointerTarget } from './drag-types';
-import { SchedulerEvent, TimeSlot } from '@mintplayer/ng-bootstrap/web-components/scheduler-core';
+import { SchedulerEvent, TimeSlot } from '@mintplayer/web-components/scheduler-core';
 
 describe('DragStateMachine', () => {
   let machine: DragStateMachine;
