+++ libs/mintplayer-ng-bootstrap/web-components/scheduler/src/index.ts
@@ -22,4 +22,4 @@ export type {
   TimeSlot,
   PreviewEvent,
   DragState,
-} from '@mintplayer/ng-bootstrap/web-components/scheduler-core';
+} from '@mintplayer/web-components/scheduler-core';
