+++ libs/mintplayer-ng-bootstrap/web-components/scheduler/src/views/day-view.ts
@@ -3,7 +3,7 @@ import {
   timelineService,
   SchedulerEventPart,
   getContrastColor,
-} from '@mintplayer/ng-bootstrap/web-components/scheduler-core';
+} from '@mintplayer/web-components/scheduler-core';
 import { BaseView, formatEventAriaLabel, isSlotInSelection } from './base-view';
 import { SchedulerState } from '../state/scheduler-state';
