+++ libs/mintplayer-ng-bootstrap/web-components/scheduler/src/views/timeline-view.ts
@@ -10,7 +10,7 @@ import {
   isResourceGroup,
   FlattenedResource,
   getContrastColor,
-} from '@mintplayer/ng-bootstrap/web-components/scheduler-core';
+} from '@mintplayer/web-components/scheduler-core';
 import { BaseView, formatEventAriaLabel, isSlotInSelection } from './base-view';
 import { SchedulerState } from '../state/scheduler-state';
