+++ libs/mintplayer-ng-bootstrap/web-components/scheduler/src/views/week-view.ts
@@ -5,7 +5,7 @@ import {
   SchedulerEventPart,
   TimeSlot,
   getContrastColor,
-} from '@mintplayer/ng-bootstrap/web-components/scheduler-core';
+} from '@mintplayer/web-components/scheduler-core';
 import { BaseView, formatEventAriaLabel, isSlotInSelection } from './base-view';
 import { SchedulerState } from '../state/scheduler-state';
