+++ libs/mintplayer-ng-bootstrap/web-components/scheduler/src/views/year-view.ts
@@ -1,7 +1,7 @@
 import {
   dateService,
   timelineService,
-} from '@mintplayer/ng-bootstrap/web-components/scheduler-core';
+} from '@mintplayer/web-components/scheduler-core';
 import { BaseView } from './base-view';
 import { SchedulerState } from '../state/scheduler-state';
