+++ libs/mintplayer-web-components/scheduler-core/src/models/drag-state.ts
