+++ libs/mintplayer-web-components/scheduler-core/src/models/event.ts
