+++ libs/mintplayer-web-components/scheduler-core/src/models/resource.ts
