+++ libs/mintplayer-web-components/scheduler-core/src/models/time-slot.ts
