+++ libs/mintplayer-web-components/scheduler-core/src/services/resource.service.ts
